refactor:implement css exports & absolute paths
