+++ src/pages/index.jsx
@@ -3,3 +3,10 @@ export { Profile } from "./profile/Profile";
 export { ProductList } from "./productList/ProductList";
 export { Wishlist } from "./wishlist/Wishlist";
 export { Cart } from "./cart/Cart";
+
+// CSS exports
+import cart from "./cart/cart.module.css";
+import filter from "./productList/filter.module.css";
+import products from "./productList/productList.module.css";
+import wish from "./wishlist/wishlist.module.css";
+export { cart, filter, products, wish };
