+++ src/pages/profile/Profile.jsx
@@ -1,4 +1,4 @@
-import { Login } from "../../components";
+import { Login } from "components";
 
 export const Profile = () => {
   return (
